Add initial support for multiple infra deployers
bifrost is currently the only way to deploy the infrastructure but
in the future other solutions will be added so we need to do some
preparation for XCI integration.

+++ xci/infra/bifrost/scripts/destroy-env.sh
@@ -36,8 +36,10 @@ fi
 # Destroy all XCI VMs if the previous operation failed
 [[ -n ${XCI_FLAVOR} ]] && \
     for vm in ${TEST_VM_NODE_NAMES}; do
-        virsh destroy $vm || true
-        virsh undefine $vm || true
+        if which virsh &>/dev/null; then
+            virsh destroy $vm || true
+            virsh undefine $vm || true
+        fi
     done
 
 service ironic-conductor stop || true
